requirements.txt
annotated-doc==0.0.4
anyio==4.13.0
asgiref==3.11.1
certifi==2026.2.25
click==8.3.2
colorama==0.4.6
Django==6.0.4
djangorestframework==3.17.1
filelock==3.25.2
fsspec==2026.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.10.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
opencv-python==4.13.0.92
packaging==26.0
pillow==12.2.0
pluggy==1.6.0
Pygments==2.20.0
pytest==9.0.3
pytest-django==4.12.0
PyYAML==6.0.3
regex==2026.4.4
rich==14.3.3
safetensors==0.7.0
setuptools==81.0.0
shellingham==1.5.4
sqlparse==0.5.5
sympy==1.14.0
timm==1.0.26
tokenizers==0.22.2
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
transformers==5.5.3
typer==0.24.1
typing_extensions==4.15.0
tzdata==2026.1
